% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots-clusters.R
\name{plot_sampled_clusters}
\alias{plot_sampled_clusters}
\title{Cluster visualization builders.}
\usage{
plot_sampled_clusters(clusters = prep_cluster_samples())
}
\arguments{
\item{clusters}{Tibble produced by \code{prep_cluster_samples()}.}
}
\value{
A \link[ggplot2:ggplot]{ggplot2::ggplot} object visualising representative clusters.
}
\description{
Plot sampled clusters of magic constants across the input range.
}
