% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform-clusters.R
\name{prep_cluster_samples}
\alias{prep_cluster_samples}
\title{Prepare sampled clusters for plotting.}
\usage{
prep_cluster_samples(n = 16384, cluster_ranges = list(c(1597622357, 1597176038), c(1598049921, 1597918602), c(1597884263, 1596910052)))
}
\arguments{
\item{n}{Number of samples per cluster.}

\item{cluster_ranges}{List of length three containing magic ranges.}
}
\value{
Tibble of sampled clusters.
}
\description{
Wraps \code{sample_clusters()} so ggplot builders can accept either raw cluster draws
or pre-computed data, bridging \code{run_cluster_pipeline()} to
\code{\link{plot_sampled_clusters}}.
}
\seealso{
\code{run_cluster_pipeline()} for regenerating the clustering datasets and
\code{\link{plot_sampled_clusters}} for visualising the prepared samples.
}
